requirements.txt
# Dependências necessárias para o Kleitos 3.0
SpeechRecognition==3.10.0
openai==0.28.1
requests==2.31.0
pygame==2.5.2
python-dotenv==1.0.0
PyAudio==0.2.13